Начал поправлять сортировку имён accession. Немного поправил test.py

+++ test.py
@@ -1,8 +1,9 @@
 #!python3.7
 import os
 
-for filename in os.listdir("./Output"):
-    os.remove(f"./Output/{filename}")
+if "Output" in os.listdir():
+    for filename in os.listdir("./Output"):
+        os.remove(f"./Output/{filename}")
 
 os.system('python3.7 PeptideSummaryAnalyzer.py "" "IDexcl.txt" ' +
           '"y" "EFRA_cont.fasta" ">=0.3" "" "default" 1 3')
@@ -14,17 +15,17 @@ for filename in os.listdir("./OutputOriginal"):
         originalFileContent = originalFile.read()
         newFileContent = newFile.read()
         line = 1
-        character = 1
+        character = 0
         for i in range(0, len(originalFileContent)):
             if i == '\n':
                 line += 1
                 character = 0
+            character += 1
             if(i == len(newFileContent) or
                originalFileContent[i] != newFileContent[i]):
                 print(f"File named {filename} not equal on line {line}" +
                       f" on character {character}!")
                 ERRORFLAG = True
                 break
-            character += 1
 
 print("Test completed {}successful!".format("un" if ERRORFLAG else ""))
